Refactored layout
