+++ pages/_document.js
@@ -1,0 +1,1 @@
+export default () => {}
